chore: refactor dropdown
